requirements.txt
torch==1.4.0
torchvision==0.5.0
pillow==8.2.0
numpy==1.19.4
foolbox==3.0.4
tensorflow-gpu==1.4.0
adversarial-robustness-toolbox
einops
git+https://github.com/fra31/auto-attack
